feat: Enable Renovate to manage node versions
We would like Renovate to open pull requests on our repositories when there are new versions of node available.

There are several files in our repository that will need changing...

* `.nvmrc`
* `.circleci/config.yml`
* `package.json`

We have already pinned the version of node used in all of our apps. With Renovate opening pull requests to upgrade node, we can release upgrades in isolation to code changes.